Changed double to BigDecimal to preserve precision for prices

+++ src/ClientCommands.scala
@@ -3,8 +3,8 @@ sealed abstract class UserCommand extends ClientCommand
 sealed abstract class AdminCommand extends ClientCommand
 
 //TODO: fix weird naming caused by trying to avoid name conflicts between client commands and server responses
-case class Buy(ticker: String, amount: Int, price: Double) extends UserCommand
-case class Sell(ticker: String, amount: Int, price: Double) extends UserCommand
+case class Buy(ticker: String, amount: Int, price: BigDecimal) extends UserCommand
+case class Sell(ticker: String, amount: Int, price: BigDecimal) extends UserCommand
 case class Cancel(orderId: Int) extends UserCommand
 case class Login(username: String, password: String) extends UserCommand
 case class Logout() extends UserCommand
@@ -15,15 +15,15 @@ case class HeldAmount(ticker: String) extends UserCommand
 
 case class AddUser(username: String, password: String) extends AdminCommand
 case class RemoveUser(username: String) extends AdminCommand
-case class ChangeUserBalance(username: String, balance: Double) extends AdminCommand
+case class ChangeUserBalance(username: String, balance: BigDecimal) extends AdminCommand
 case class SetUserAssets(username: String, ticker: String, amount: Int) extends AdminCommand
 
 object ClientCommand {
   def parse(command: String) = {
     val tokens = command.trim.split(' ')
     if (tokens.length == 0) None else try tokens(0) match {
-      case "BUY"        => Buy(tokens(1), tokens(2).toInt, tokens(3).toDouble)
-      case "SELL"       => Sell(tokens(1), tokens(2).toInt, tokens(3).toDouble)
+      case "BUY"        => Buy(tokens(1), tokens(2).toInt, BigDecimal(tokens(3)))
+      case "SELL"       => Sell(tokens(1), tokens(2).toInt, BigDecimal(tokens(3)))
       case "CANCEL"     => Cancel(tokens(1).toInt)
       case "LOGIN"      => Login(tokens(1), tokens(2))
       case "LOGOUT"     => Logout()
@@ -33,7 +33,7 @@ object ClientCommand {
       case "ASSETS"     => HeldAmount(tokens(1))
       case "ADDUSER"    => AddUser(tokens(1), tokens(2))
       case "REMOVEUSER" => RemoveUser(tokens(1))
-      case "SETBALANCE" => ChangeUserBalance(tokens(1), tokens(2).toDouble)
+      case "SETBALANCE" => ChangeUserBalance(tokens(1), BigDecimal(tokens(2)))
       case "SETASSETS"  => SetUserAssets(tokens(1), tokens(2), tokens(3).toInt)
       case _            => None
     } catch {
